move String methods to separate file
